python_implementation/requirements.txt
fal_client==0.4.0
gradio==4.31.5
pillow